+++ packages/quantic/force-app/main/default/lwc/quanticUtils/quanticUtils.js
@@ -4,6 +4,8 @@ import dayPattern from '@salesforce/label/c.quantic_DatePatternDay';
 import monthPattern from '@salesforce/label/c.quantic_DatePatternMonth';
 import yearPattern from '@salesforce/label/c.quantic_DatePatternYear';
 
+/** @typedef {import("coveo").Result} Result */
+
 export class Debouncer {
   _timeout;
 
@@ -687,4 +689,45 @@ export function isParentOf(element, targetElement) {
     (acc, val) => acc || isParentOf(val, targetElement),
     false
   );
+}
+
+/**
+ * Copies text to clipboard using the Clipboard API.
+ * https://developer.mozilla.org/en-US/docs/Web/API/Clipboard
+ * @param {string} text
+ */
+export async function copyToClipboard(text) {
+  await navigator.clipboard.writeText(text);
+}
+
+/**
+ * Read the value of a given key from an object.
+ * @param {object} object 
+ * @param {string} key 
+ */
+export function readFromObject(object, key) {
+  const firstPeriodIndex = key.indexOf('.');
+  if (object && firstPeriodIndex !== -1) {
+    let newKey = key.substring(firstPeriodIndex + 1);
+    key = key.substring(0, firstPeriodIndex);
+    return readFromObject(object[key], newKey);
+  }
+  return object ? object[key] : undefined;
+}
+
+/**
+ * Generates a text from a result based on a given template.
+ * @param {string} template 
+ * @param {Result} result 
+ * @returns {string}
+ */
+export function buildTemplateTextFromResult(template, result) {
+  if (!template) {
+    return '';
+  }
+  return template.replace(/\$\{(.*?)\}/g, (value) => {
+    const key = value.substring(2, value.length - 1);
+    const newValue = readFromObject(result, key);
+    return newValue || value;
+  });
 }