+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.126.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.125.1...@coveo/atomic@1.126.0) (2022-12-09)
+
+### Features
+
+- **atomic:** add support for min and max for atomic-timeframe-facet ([#2596](https://github.com/coveo/ui-kit/issues/2596)) ([0f34acc](https://github.com/coveo/ui-kit/commit/0f34acc1b15c1bcbf27202f4586ae77cb2a76ed5))
+
 ## [1.125.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.125.0...@coveo/atomic@1.125.1) (2022-12-06)
 
 ### Bug Fixes